add support for system properties in configuration
Change the configuration parser of the MQTT parser, so we could use
system properties in the value of a configuration key.
The configuration file that the OSGi bundle is using points normally to
${moquette.path}/config/moquette.conf.
So we could use a configuration file that contains an entry e.g.
persistent_store ${moquette.path}/storage/moquette_store.mapdb
So the real location of the configuration and the dabase storage could
be changed for different environments without touching anything else but
a system property.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>