Improving th plugin manager's code & Adding connect/disconnect plugin.
Making the call for inject static in the Server class.
Adding a plugin that will be called when a user connect and disconnect.
The default plugin is enabled by default and it doesn't do anything.

To add a new con/dis plugin, create a class and let it inheret from
IConDisPlugin or extens the class AbstractConDisPlugin. Then, set that
class in the IConDisService entry in the plugins file, which exists in
the broker module under config folder.