Introduced test to verify that connectionLost remove the new client session (#557)
This PR add the link MQTTConnection -> Session so now the two entities are doubly linked. This permit the removal of session from session registry iff the session it's removing is the only one active in the session registry. This also avoid the continuous access to SessionRegistry from the MQTTConnection to retrieve the session linked to the MQTTConnection.

PR #557