Fixed ServerLowLevel tests not running
JUnit only runs tests in classes that have a name that ends in Test,
the class in question had a name that ended in Tests, with an added s.
As a result, the tests in this class were never executed. This renames
the class to not have the added s.