Fix #423 resent messag not added to timeout queue (#582)
* Resent messages are not added to inflightTimeouts.

Messages that are re-sent were not added to the timeout queue, as a
result, if the re-sent failed as well, the message was not resent again,
instead remaining on the inflightWindow for ever. By re-adding it to the
inflightTimeouts, the message is "kept alive" until sending succeeds.

* Fix possible Race condition: change contains + get to get + null check.

Between the contains and get calls, another thread could steal the
value. This also saves one hash lookup.

Co-authored-by: andsel <selva.andre@gmail.com>