Give precedence of PUBREC processing instead of the queued
Changes the way PUBREC is managed, instead of return the lease on inflight and request it again, the lease is kept to complete the flow instead of give precedence to enqueued messages. This is to in principle to "complete the work we are doing" instead of open as many flows as we can.