Introduces sessions event processors to segregate changes to a session in one single thread, simplifying concurrency and code(#631)
Introduces the concept of sticky sessions, all the change operations to a session is done always by the same thread, so this simplifies the concurrency access, segregating the access to only one thread. The same thread can handle multiple sessions.

Every session event processor is an event loop, that peeks commands out of a queue and executes them.
This means that there is an isolation queue between the IO threads that executes the lower layer of protocol and the event loops that mutate and managed the upper logic, mainly bound to Session's related structures.

The length of those queue can be configured with option session_queue_size, by default is has 1024 slots, but it could be that if there is an heavy publisher, the queue becomes full and this case the only thing to do is log an error. So it could happen that for a SUB command no SUBACK is sent (the same is valid for PUB) if a queue is full. In MQTT 3.1.1 there is no timeout specification for missed SUBACK, so the client will not send a second SUB (dup=1) and we potentially could loose SUBscriptions (and PUBlishes).

Co-authored-by: Hylke van der Schaaf <hylke.vds@gmail.com>