Persisted unsafe queues implementation (#691)
Create the implementation of segmented queues. The data is stored in mmapped memory pages, where each page is divided in segments. Each queue own a list of segments. The access to the each segment is intended to be done by the same thread, so no serialization is needed. The only critical section, guarded by a lock, is the interaction to require or release segment.

General design discussion at https://github.com/moquette-io/moquette/discussions/663