Sets clientId into Channel's attribute immediately after processing the connect
MQTT 3.1.1 specification, MQTT-3.1.4-5 says that Clients are allowed to send further Control Packets immediately after sending a CONNECT Packet.
In this case a PUBLISH is could be sent immediately after the CONNECT, so it's processed serially by the same IO thread.
The PUBLISH needs to accees the clientID attribute of the Channel, so it needs be set before the CONNACK in write&flush.