Bubble session loops excpetions
When a session loop the execption is simply logged and lost, so the error status is not reported to the broker.
This PR interrupt the session loop on an exception, saving the error in a local variable.

When the PostOffice terminates it re-throw the first (and last) exception that was reached during session loop
execution.