Create a DSL to ease the creation of a Server instance (#761)
Create a fluent class (FluentConfig) to instantiate, configure and run the server in one liner.
Instead of creating an instance of IConfig and invoking repeatedly put methods with some constants that has to be fetched from a pool.

Implementation of a DSL to ease the configuration of the Server without boring use of Properties and BrokerConstants. This simplify the embedding of the Server in other projects.