Handle 0 sized collector batches in PostOffice (#777)
This change is attempting to resolve #750 where the call to retain is 0 and therefore countBatches must have returned 0. I know that this change will absolutely avoid the error, however, I'm not clear on why the number of batches would be 0 while this code is running. In this change, I simply treat it exactly the same as the above case where there are no subscriptions (and thus no work to do).