Initial implementation of shared subscriptions (#796)
Contains a series of updates to the broker internal structures to sustain shared subscription registering and dispatching of publish messages.

In CTrie's CNode add a separate shared subscriptions data structure to contains the shared subscriptions list related to share name for a given topic filter. It's implemented as a map keyed by share name which maps to a list of shared subscriptions.
Updated the CTrie's insertion method to use a data class SubscriptionRequest to capture the shared and non shared subscription actions.
Implemented test cases to cover the various MQTT spec requirements as integrations tests, and adapted utility classes (like the low level Client to permit the verification).