Payload format verification  and content type properties handling (#826)
Updates the processing of publish messages (for all QoS) to verify the payload to be encoded in UTF-8 when property `payload format indicator` is set. If the validation fails, return an acknowledgement message (PUBACK for QoS1 and PUBREC for QoS2, and DISCONNECT for QoS0) with an appropriate return code.
Manage also `content_type` property.